Handle messages from CAN
